Runned checkStyle and formated code allong with changing the View package to view (to comply with checkStyle).
