+++ src/view/BoardSpaces/CardSpace.java
@@ -1,4 +1,4 @@
-package View.BoardSpaces;
+package view.BoardSpaces;
 
 import java.awt.Color;
 import java.awt.Dimension;
