+++ src/view/BoardSpaces/FreeParkingSpace.java
@@ -1,4 +1,4 @@
-package View.BoardSpaces;
+package view.BoardSpaces;
 
 import java.awt.*;
 import java.io.File;
@@ -8,13 +8,13 @@ import javax.swing.*;
 import javax.swing.border.Border;
 
 /**********************************************************************
- * The GoSpace class creates a JPanel that resembles a Go Square from
- * the game of Monopoly.
+ * The FreeParkingSpace class creates a JPanel that resembles a Free
+ * Parking Square from the game of Monopoly.
  *
- * @author David Baas
+ * @author Dustin Ecker
  * @version 3/26/2018
  *********************************************************************/
-public class GoSpace extends Spaces {
+public class FreeParkingSpace extends Spaces {
 
   /**
    * A static int constant to track the length of an edge in the square.
@@ -22,40 +22,38 @@ public class GoSpace extends Spaces {
   private static final int LENGTH = 100;
 
   /**
-   * An image to hold the image of the Go icon.
+   * An image to hold the image of the Free Parking icon.
    */
-  private static Image goImg;
-
+  private static Image freeParkingImg;
 
   /********************************************************************
    * The constructor initializes the image.
    *
    * @throws IOException If there is an error in loading the image.
    *******************************************************************/
-  public GoSpace() throws IOException {
+  public FreeParkingSpace() throws IOException {
     super(true);
-
-    goImg = ImageIO.read(new File("res/goImg.gif"));
-
+    freeParkingImg = ImageIO.read(new File("res/freeParkingImg.png"));
     Border blackLine = BorderFactory.createLineBorder(Color.BLACK);
     setBorder(blackLine);
+
   }
 
-  /*******************************************************************
-   * The paintComponent draws the Go Image onto the JPanel.
+  /********************************************************************
+   * The paintComponent draws Free Parking image onto the JPanel.
    *
    * @param g Graphics component of this panel.
-   ******************************************************************/
+   ********************************************************************/
   public void paintComponent(Graphics g) {
     super.paintComponent(g);
-    g.drawImage(goImg, 0, 0, LENGTH, LENGTH, null);
+    g.drawImage(freeParkingImg, 0, 0, LENGTH, LENGTH, null);
     drawTokens(g);
   }
 
   /********************************************************************
    * The getPreferredSize method is used to Lock the size of the Panel
    * to the correct size.
-   * @return The dimensions of the GoSquare.
+   * @return The dimensions of the FreeParkingSquare.
    *******************************************************************/
   @Override
   public Dimension getPreferredSize() {
@@ -65,7 +63,7 @@ public class GoSpace extends Spaces {
   /********************************************************************
    * The getMinimumSize method refers to the getPreferredSize method
    * to lock the size of the panel.
-   * @return getPreferredSize the dimensions of the GoSquare.
+   * @return getPreferredSize the dimensions of the FreeParkingSquare.
    *******************************************************************/
   @Override
   public Dimension getMinimumSize() {
@@ -75,7 +73,7 @@ public class GoSpace extends Spaces {
   /********************************************************************
    * The getMaximumSize method refers to the getPreferredSize method
    * to lock the size of the panel.
-   * @return getPreferredSize the dimensions of the GoSquare.
+   * @return getPreferredSize the dimensions of the FreeParkingSquare.
    *******************************************************************/
   @Override
   public Dimension getMaximumSize() {
