+++ src/view/BoardSpaces/GoSpace.java
@@ -1,4 +1,4 @@
-package View.BoardSpaces;
+package view.BoardSpaces;
 
 import java.awt.*;
 import java.io.File;
@@ -8,13 +8,13 @@ import javax.swing.*;
 import javax.swing.border.Border;
 
 /**********************************************************************
- * The FreeParkingSpace class creates a JPanel that resembles a Free
- * Parking Square from the game of Monopoly.
+ * The GoSpace class creates a JPanel that resembles a Go Square from
+ * the game of Monopoly.
  *
- * @author Dustin Ecker
+ * @author David Baas
  * @version 3/26/2018
  *********************************************************************/
-public class FreeParkingSpace extends Spaces {
+public class GoSpace extends Spaces {
 
   /**
    * A static int constant to track the length of an edge in the square.
@@ -22,38 +22,40 @@ public class FreeParkingSpace extends Spaces {
   private static final int LENGTH = 100;
 
   /**
-   * An image to hold the image of the Free Parking icon.
+   * An image to hold the image of the Go icon.
    */
-  private static Image freeParkingImg;
+  private static Image goImg;
+
 
   /********************************************************************
    * The constructor initializes the image.
    *
    * @throws IOException If there is an error in loading the image.
    *******************************************************************/
-  public FreeParkingSpace() throws IOException {
+  public GoSpace() throws IOException {
     super(true);
-    freeParkingImg = ImageIO.read(new File("res/freeParkingImg.png"));
+
+    goImg = ImageIO.read(new File("res/goImg.gif"));
+
     Border blackLine = BorderFactory.createLineBorder(Color.BLACK);
     setBorder(blackLine);
-
   }
 
-  /********************************************************************
-   * The paintComponent draws Free Parking image onto the JPanel.
+  /*******************************************************************
+   * The paintComponent draws the Go Image onto the JPanel.
    *
    * @param g Graphics component of this panel.
-   ********************************************************************/
+   ******************************************************************/
   public void paintComponent(Graphics g) {
     super.paintComponent(g);
-    g.drawImage(freeParkingImg, 0, 0, LENGTH, LENGTH, null);
+    g.drawImage(goImg, 0, 0, LENGTH, LENGTH, null);
     drawTokens(g);
   }
 
   /********************************************************************
    * The getPreferredSize method is used to Lock the size of the Panel
    * to the correct size.
-   * @return The dimensions of the FreeParkingSquare.
+   * @return The dimensions of the GoSquare.
    *******************************************************************/
   @Override
   public Dimension getPreferredSize() {
@@ -63,7 +65,7 @@ public class FreeParkingSpace extends Spaces {
   /********************************************************************
    * The getMinimumSize method refers to the getPreferredSize method
    * to lock the size of the panel.
-   * @return getPreferredSize the dimensions of the FreeParkingSquare.
+   * @return getPreferredSize the dimensions of the GoSquare.
    *******************************************************************/
   @Override
   public Dimension getMinimumSize() {
@@ -73,7 +75,7 @@ public class FreeParkingSpace extends Spaces {
   /********************************************************************
    * The getMaximumSize method refers to the getPreferredSize method
    * to lock the size of the panel.
-   * @return getPreferredSize the dimensions of the FreeParkingSquare.
+   * @return getPreferredSize the dimensions of the GoSquare.
    *******************************************************************/
   @Override
   public Dimension getMaximumSize() {
