+++ src/view/BoardSpaces/PropertySpace.java
@@ -1,4 +1,4 @@
-package View.BoardSpaces;
+package view.BoardSpaces;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
