+++ src/view/GamePanel.java
@@ -1,16 +1,16 @@
-package View;
+package view;
 
-import View.BoardSpaces.CardSpace;
-import View.BoardSpaces.FreeParkingSpace;
-import View.BoardSpaces.GoSpace;
-import View.BoardSpaces.GoToJailSpace;
-import View.BoardSpaces.JailSpace;
-import View.BoardSpaces.PropertySpace;
-import View.BoardSpaces.PropertySpace.Position;
-import View.BoardSpaces.RailRoadSpace;
-import View.BoardSpaces.Spaces;
-import View.BoardSpaces.TaxSpace;
-import View.BoardSpaces.UtilitiesSpace;
+import view.BoardSpaces.CardSpace;
+import view.BoardSpaces.FreeParkingSpace;
+import view.BoardSpaces.GoSpace;
+import view.BoardSpaces.GoToJailSpace;
+import view.BoardSpaces.JailSpace;
+import view.BoardSpaces.PropertySpace;
+import view.BoardSpaces.PropertySpace.Position;
+import view.BoardSpaces.RailRoadSpace;
+import view.BoardSpaces.Spaces;
+import view.BoardSpaces.TaxSpace;
+import view.BoardSpaces.UtilitiesSpace;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.GridBagConstraints;
